search/requirements.txt
kafka-python
psycopg2
google-api-python-client
stringcase